Tier 3: Fix 4 data integrity bugs (M2, M7, M8, C2)
- M2: Earnings gate rewritten to use yfinance (Alpaca API doesn't support
  earnings). Fail-open preserved, 24h cache, optional dependency.
- M7: signal_logs.step_stopped_at column type changed from INTEGER to REAL
  to support fractional step numbers (e.g., 8.7, 9.5)
- M8: Profile delete now cleans up training_queue and system_state rows
  to prevent orphaned data accumulation
- C2: Walk-forward backtest passes train_end date to train_model() via new
  data_end_override parameter. Each window now trains on different data.

+++ options-bot/backend/database.py
@@ -105,7 +105,7 @@ CREATE TABLE IF NOT EXISTS signal_logs (
     underlying_price REAL,
     predicted_return REAL,
     predictor_type TEXT,
-    step_stopped_at INTEGER,
+    step_stopped_at REAL,
     stop_reason TEXT,
     entered INTEGER DEFAULT 0,
     trade_id TEXT
